Filter which props are passed into the underlying span
Previously the JSX splat syntax was used to pass `this.props` in its entirety into the underlying `<span>` component. This resulted in invalid HTML markup because the `name` prop was rendered in the markup, and according to the W3 HTML checker it's invalid to have a `name` prop on a `<span>` in most contexts.

So instead, we split out the FontAwesome-specific props from the rest of them, and pass only the non-FontAwesome props down to the `<span>`.

Test Plan: Added a regression test